requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
